+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 2.0.2
 * `_join` previously discarded quer parameters, etc.
+* Allow any `Map<String, dynamic>` as body, not just `Map<String, String>`.
 
 # 2.0.1
 * Change `BaseAngelClient` constructor to accept `dynamic` instead of `String` for `baseUrl.